+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-## [Unreleased]
-### Fixes
+## [Unreleased] - Improvements
 - removed ns3: from all responses
+- added a escaping of sender and receiver values
 
 ## [1.0.5] - Updates
 - Added length to measurements